+++ src/app/view-config-upgraders.js
@@ -334,3 +334,18 @@ export function upgradeFrom1_0_8(config) {
     version: '1.0.9',
   };
 }
+
+// TODO
+
+// Added in version 1.0.13:
+// - Adds the property `coordinationValues` for
+// view config file definitions but is not yet
+// used to do file matching/lookups.
+export function upgradeFrom1_0_12(config) {
+  const newConfig = cloneDeep(config);
+
+  return {
+    ...newConfig,
+    version: '1.0.13',
+  };
+}