+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 
 ### Added
+- Added the property `coordinationValues` for view config file definitions but is not yet used to do file matching/lookups.
 
 ### Changed